+++ dao/src/main/java/org/n52/series/db/da/FeatureRepository.java
@@ -29,119 +29,46 @@
 
 package org.n52.series.db.da;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
 import org.hibernate.Session;
-import org.n52.io.request.IoParameters;
 import org.n52.io.response.FeatureOutput;
-import org.n52.series.db.DataAccessException;
-import org.n52.series.db.beans.DescribableEntity;
 import org.n52.series.db.beans.FeatureEntity;
 import org.n52.series.db.beans.parameter.Parameter;
 import org.n52.series.db.dao.DbQuery;
 import org.n52.series.db.dao.FeatureDao;
+import org.n52.series.db.dao.SearchableDao;
 import org.n52.series.spi.search.FeatureSearchResult;
 import org.n52.series.spi.search.SearchResult;
-import org.n52.web.exception.ResourceNotFoundException;
 
 public class FeatureRepository extends HierarchicalParameterRepository<FeatureEntity, FeatureOutput> {
 
     @Override
-    public boolean exists(String id, DbQuery parameters) throws DataAccessException {
-        Session session = getSession();
-        try {
-            FeatureDao dao = createDao(session);
-            return dao.hasInstance(parseId(id), parameters, FeatureEntity.class);
-        } finally {
-            returnSession(session);
-        }
-    }
-    @Override
-    public Collection<SearchResult> searchFor(IoParameters parameters) {
-        Session session = getSession();
-        try {
-            FeatureDao featureDao = createDao(session);
-            DbQuery query = dbQueryFactory.createFrom(parameters);
-            List<FeatureEntity> found = featureDao.find(query);
-            return convertToSearchResults(found, query);
-        } finally {
-            returnSession(session);
-        }
+    protected FeatureOutput prepareEmptyParameterOutput(FeatureEntity entity) {
+        return new FeatureOutput();
     }
 
     @Override
-    public List<SearchResult> convertToSearchResults(List< ? extends DescribableEntity> found, DbQuery query) {
-        String locale = query.getLocale();
-        String hrefBase = urlHelper.getFeaturesHrefBaseUrl(query.getHrefBase());
-        List<SearchResult> results = new ArrayList<>();
-        for (DescribableEntity searchResult : found) {
-            String pkid = searchResult.getPkid()
-                                      .toString();
-            String label = searchResult.getLabelFrom(locale);
-            results.add(new FeatureSearchResult(pkid, label, hrefBase));
-        }
-        return results;
+    protected SearchResult createEmptySearchResult(String id, String label, String baseUrl) {
+        return new FeatureSearchResult(id, label, baseUrl);
     }
 
     @Override
-    public List<FeatureOutput> getAllCondensed(DbQuery parameters) throws DataAccessException {
-        Session session = getSession();
-        try {
-            return getAllCondensed(parameters, session);
-        } finally {
-            returnSession(session);
-        }
+    protected String createHref(String hrefBase) {
+        return urlHelper.getFeaturesHrefBaseUrl(hrefBase);
     }
 
     @Override
-    public List<FeatureOutput> getAllCondensed(DbQuery parameters, Session session) throws DataAccessException {
-        return createCondensed(getAllInstances(parameters, session), parameters);
+    protected FeatureDao createDao(Session session) {
+        return new FeatureDao(session);
     }
 
     @Override
-    public List<FeatureOutput> getAllExpanded(DbQuery parameters) throws DataAccessException {
-        Session session = getSession();
-        try {
-            return getAllExpanded(parameters, session);
-        } finally {
-            returnSession(session);
-        }
+    protected SearchableDao<FeatureEntity> createSearchableDao(Session session) {
+        return new FeatureDao(session);
     }
 
     @Override
-    public List<FeatureOutput> getAllExpanded(DbQuery parameters, Session session) throws DataAccessException {
-        return createExpanded(getAllInstances(parameters, session), parameters);
-    }
-
-    private List<FeatureEntity> getAllInstances(DbQuery parameters, Session session) throws DataAccessException {
-        return createDao(session).getAllInstances(parameters);
-    }
-
-    @Override
-    public FeatureOutput getInstance(String id, DbQuery parameters) throws DataAccessException {
-        Session session = getSession();
-        try {
-            return getInstance(id, parameters, session);
-        } finally {
-            returnSession(session);
-        }
-    }
-
-    @Override
-    public FeatureOutput getInstance(String id, DbQuery parameters, Session session) throws DataAccessException {
-        FeatureDao featureDao = createDao(session);
-        FeatureEntity result = featureDao.getInstance(parseId(id), parameters);
-        if (result == null) {
-            throw new ResourceNotFoundException("Resource with id '" + id + "' could not be found.");
-        }
-        return createExpanded(result, parameters);
-    }
-
-    @Override
-    protected FeatureOutput createExpanded(FeatureEntity entity, DbQuery parameters) throws DataAccessException {
-        FeatureOutput result = createCondensed(entity, parameters);
+    protected FeatureOutput createExpanded(FeatureEntity entity, DbQuery parameters, Session session) {
+        FeatureOutput result = createCondensed(entity, parameters, session);
         if (parameters.getHrefBase() != null) {
             result.setService(getCondensedExtendedService(getServiceEntity(entity), parameters));
         } else {
@@ -153,22 +80,6 @@ public class FeatureRepository extends HierarchicalParameterRepository<FeatureEn
             }
         }
         return result;
-    }
-
-    @Override
-    protected FeatureOutput createCondensed(FeatureEntity entity, DbQuery parameters) {
-        FeatureOutput result = new FeatureOutput();
-        result.setId(Long.toString(entity.getPkid()));
-        result.setLabel(entity.getLabelFrom(parameters.getLocale()));
-        result.setDomainId(entity.getDomainId());
-        checkForHref(result, parameters);
-        return result;
-    }
-
-    private void checkForHref(FeatureOutput result, DbQuery parameters) {
-        if (parameters.getHrefBase() != null) {
-            result.setHrefBase(urlHelper.getFeaturesHrefBaseUrl(parameters.getHrefBase()));
-        }
     }
 
 }