+++ dao/src/main/java/org/n52/series/db/da/OutputAssembler.java
@@ -35,7 +35,6 @@ import java.util.List;
 import org.hibernate.Session;
 import org.n52.io.request.IoParameters;
 import org.n52.series.db.DataAccessException;
-import org.n52.series.db.beans.DescribableEntity;
 import org.n52.series.db.dao.DbQuery;
 import org.n52.series.spi.search.SearchResult;
 
@@ -54,8 +53,6 @@ public interface OutputAssembler<T> {
     T getInstance(String id, DbQuery parameters, Session session) throws DataAccessException;
 
     Collection<SearchResult> searchFor(IoParameters parameters);
-
-    List<SearchResult> convertToSearchResults(List< ? extends DescribableEntity> found, DbQuery query);
 
     boolean exists(String id, DbQuery query) throws DataAccessException;