+++ dao/src/main/java/org/n52/series/db/da/SearchableRepository.java
@@ -1,0 +1,11 @@
+package org.n52.series.db.da;
+
+import java.util.Collection;
+
+import org.n52.io.request.IoParameters;
+import org.n52.series.spi.search.SearchResult;
+
+public interface SearchableRepository {
+
+    Collection<SearchResult> searchFor(IoParameters paramters);
+}